implement "attack star base" button

+++ scripts/events/AbandonedVesselMapEvent.cs
@@ -44,7 +44,7 @@ public class AbandonedVesselMapEvent : AbstractMapEvent {
                             },
                             new Effect{
                                 kind = EffectKind.AddFuel,
-                                value = (float)fuel,
+                                value = fuel,
                             },
                         },
                     };
